Define a generic Component.
There is a some basic functionality that can be moved to up the chain of
commands.

I also make paired down the test file to make it faster to run the View
file dependant stuff.

And I got rid of the old Eval.xhtml... way to big.

Finally I made it so that slow file dependant tests are optional so I've
changed our casperTest.sh script to reflect that change.

Issue 1296